Create new version that supports fragments
Fragment spreads inside of selection sets are hard to model with Flow
since there is no equivalent in Flow. On the first level they can be
done with type intersections but when the types are nested the
behaviours differ. To make this possible I created a function that
merges two object type definitions into a single one. This way we can
hardcode the fragment behaviour in the type definitions.